back/application-structure: configure application structure  - configure models : User, Post, Subscription, Topic  - configure TopicRepository  - configure model mapper : add bean configuration to    configuration/AppConfig  - configure spécific mappers : UserMapper and TopicMapper in mapper    package  - add Dtos to dto package

+++ back/src/main/java/com/jerem/mdd/service/SubscriptionService.java
@@ -3,13 +3,14 @@ package com.jerem.mdd.service;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+
 import com.jerem.mdd.model.Subscription;
 
 @Service
 public class SubscriptionService {
 
 
-    public List<Subscription> getSubscriptionsByUser(Integer id) {
+    public List<Subscription> getSubscriptionsByUser(Long id) {
         // TODO return the subscription of the user
         return null;
     }
